back_end/pyproject.toml
[project]
name = "magnus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.124.0",
    "httpx>=0.28.1",
    "pywheels>=0.8.3.2",
    "ruamel-yaml>=0.18.16",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "uvicorn>=0.38.0",
]
